moved realm.json to root level + removed unneeded variable Authentication in LogIn.jsx + fixed <Link> warning is marked as required in , but its value is  by replacing w/ <a/>
