+++ electron/realm.json
@@ -1,0 +1,3 @@
+{
+  "appId": "electron-sample-app-nzhsm"
+}
