--- electron/src/realm.json
@@ -1,3 +1,0 @@
-{
-  "appId": "myapp-zufnj"
-}
